Add a lock util
Realized that messages could be sent over the same stream while we're
waiting for a response for the previous message. This is not good. Made
a lock util to help with that.

I see this happening a lot once we start trying to do completion type
items, so that's where I came up with this part.